4.28. Refatorando a exclusão de cozinhas para usar domain services
Criado no servico do dominio o metodo de excluir e tratando o caso de
sucesso, nao encontrado (404) e conflito com banco de dados (409)
Criado 2 excecoes de negocios pra encapsular e identificar melhor a
ocorrencia de erro

+++ algafood-api/src/main/java/com/algaworks/algafood/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,8 +36,11 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 	
 	@Override
 	@Transactional
-	public void remover(Cozinha cozinha) {
-		cozinha = buscarPorId(cozinha.getId());
+	public void remover(Long id) {
+		Cozinha cozinha = buscarPorId(id);
+		if (cozinha == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
 		entityManager.remove(cozinha);
 	}
 }
